requirements.txt
pillow==9.0.1
selenium==3.141.0
pytest==8.0.0
